starts UI testing
Worked around race condition by waiting till number of tables reaches the
expected count of 1.

There is no baked in long-press action, but adding one yourself is quite
simple.

Because the cell UI element's value is just a plain string, we have no
way to see that it's been crossed out, so no way to detect
whether it's finished or not from the UI tests.
This stumps our testing for a bit, but it also points out an accessibility
issue that was hiding in plain sight: How will an unsighted user know whether
a todo item is finished or not?