groups accessibility logic
Unlike logic tests, UI tests do not execute within the app executable itself,
so they don't get to freely use app code, not even with an `@testable import`.

In order to reuse accessibility label details from the UI test code,
we need to add the file containing that app code and any dependencies
it requires to build to the UI test target.
If we left the logic in TodoCellView, we'd thus be forced to build
Todo.swift as part of the UI test target as well, even though the UI tests
never touch it.

Shifting code shared between the app and the UI test target
to a leaf file avoids pulling in dead code.
It also removes accessibility string-formatting details from the
list of responsibilities assigned to the TodoCellView.