increases simulated long-press duration
0.5 seconds was failing to trigger the long-press gesture recognizer.
Just slightly longer appears to fix the issue.
There's a tension between wanting to make sure it gets triggered,
so as to avoid a flickering test, and wanting to move on ASAP once it
has been triggered, so as to minimize how long it takes to run all the
UI tests.

Since all the long presses use the same duration, if we find it needs to be
longer later to avoid spurious test failures, it will be no problem to
adjust it then.